Financeiro  Matricula e Mensalidade

+++ AcademiaDanca.Dominio/Contexto/Comandos/FinanceiroComando/Entrada/MatricularComando.cs
@@ -15,7 +15,7 @@ namespace AcademiaDanca.IO.Dominio.Contexto.Comandos.FinanceiroComando.Entrada
         public decimal PercentualDesconto { get; set; }
         public decimal ValorDesconto { get; set; }
         public decimal ValorMaricula { get; set; }
-        public decimal ValorContrato { get; set; }
+        public double ValorContrato { get; set; }
         public int DiaVencimento { get; set; }
         public int TotalParcelas { get; set; }
         public int Ano { get; set; }
